requirements.txt
nonebot2[websockets]
nonebot-adapter-onebot
redis
